+++ bin/ci/phpunit.inc.php
@@ -6,29 +6,19 @@ use Steevanb\ParallelProcess\{
     Process\Process,
     Process\ProcessArray
 };
-use steevanb\PhpTypedArray\ScalarArray\StringArray;
+use Steevanb\PhpTypedArray\ScalarArray\StringArray;
 
 function createPhpunitProcesses(string $phpVersion = null, string $symfonyVersion = null): ProcessArray
 {
-    $phpVersions = new StringArray(is_string($phpVersion) ? [$phpVersion] : ['7.1', '7.2', '7.3', '7.4', '8.0']);
+    $phpVersions = new StringArray(is_string($phpVersion) ? [$phpVersion] : ['8.1', '8.2']);
     $symfonyVersions = new StringArray(
-        is_string($symfonyVersion) ? [$symfonyVersion] : ['4.4', '5.0', '5.1', '5.2', '5.3']
+        is_string($symfonyVersion) ? [$symfonyVersion] : ['6.1', '6.2']
     );
 
     $return = new ProcessArray();
     foreach ($phpVersions as $loopPhpVersion) {
         foreach ($symfonyVersions as $loopSymfonyVersion) {
-            if (
-                (
-                    in_array($loopPhpVersion, ['7.2', '7.3', '7.4', '8.0'])
-                    && in_array($loopSymfonyVersion, ['4.4', '5.0', '5.1', '5.2', '5.3'])
-                ) || (
-                    $loopPhpVersion === '7.1'
-                    && $loopSymfonyVersion === '4.4'
-                )
-            ) {
-                $return[] = createPhpunitProcess($loopPhpVersion, $loopSymfonyVersion);
-            }
+            $return[] = createPhpunitProcess($loopPhpVersion, $loopSymfonyVersion);
         }
     }