+++ bridge/Symfony/PhpTypedArrayBundle.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace steevanb\PhpTypedArray\Bridge\Symfony;
+namespace Steevanb\PhpTypedArray\Bridge\Symfony;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;