+++ src/ObjectArray/UnicodeStringArray.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace steevanb\PhpTypedArray\ObjectArray;
+namespace Steevanb\PhpTypedArray\ObjectArray;
 
 use Symfony\Component\String\UnicodeString;
 
@@ -14,7 +14,7 @@ class UnicodeStringArray extends ObjectArray
         parent::__construct($values, UnicodeString::class);
     }
 
-    public function merge(UnicodeStringArray $typedArray): self
+    public function merge(UnicodeStringArray $typedArray): static
     {
         parent::doMerge($typedArray);