+++ tests/Unit/AbstractTypedArray/TypedArray.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace steevanb\PhpTypedArray\Tests\Unit\AbstractTypedArray;
+namespace Steevanb\PhpTypedArray\Tests\Unit\AbstractTypedArray;
 
-use steevanb\PhpTypedArray\AbstractTypedArray;
+use Steevanb\PhpTypedArray\AbstractTypedArray;
 
 final class TypedArray extends AbstractTypedArray
 {