+++ tests/Unit/AbstractTypedArray/ValudAlreadyExistTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace steevanb\PhpTypedArray\Tests\Unit\AbstractTypedArray;
+namespace Steevanb\PhpTypedArray\Tests\Unit\AbstractTypedArray;
 
 use PHPUnit\Framework\TestCase;
-use steevanb\PhpTypedArray\Exception\ValueAlreadyExistException;
+use Steevanb\PhpTypedArray\Exception\ValueAlreadyExistException;
 
 final class ValudAlreadyExistTest extends TestCase
 {