+++ tests/Unit/ObjectArray/TestObject.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace steevanb\PhpTypedArray\Tests\Unit\ObjectArray;
+namespace Steevanb\PhpTypedArray\Tests\Unit\ObjectArray;
 
 final class TestObject
 {